fix: arrange fileUploader arborescence, add small features
