+++ packages/components/music/UploadMusicModal/UploadTrack.tsx
@@ -38,8 +38,8 @@ import { BrandText } from "../../BrandText";
 import { EditableAudioPreview } from "../../FilePreview/EditableAudioPreview";
 import { SVG } from "../../SVG";
 import { PrimaryButton } from "../../buttons/PrimaryButton";
-import { FileUploader } from "../../fileUploader";
 import { TextInputCustom } from "../../inputs/TextInputCustom";
+import { FileUploader } from "../../inputs/fileUploader";
 import { FeedPostingProgressBar } from "../../loaders/FeedPostingProgressBar";
 import { FeedFeeText } from "../../socialFeed/FeedFeeText";
 import { SpacerColumn, SpacerRow } from "../../spacer";
