+++ packages/components/video/UploadVideoModal.tsx
@@ -44,8 +44,8 @@ import { OptimizedImage } from "../OptimizedImage";
 import { SVG } from "../SVG";
 import { CustomPressable } from "../buttons/CustomPressable";
 import { PrimaryButton } from "../buttons/PrimaryButton";
-import { FileUploader } from "../fileUploader";
 import { TextInputCustom } from "../inputs/TextInputCustom";
+import { FileUploader } from "../inputs/fileUploader";
 import { FeedPostingProgressBar } from "../loaders/FeedPostingProgressBar";
 import { SelectAudioVideo } from "../mini/SelectAudioVideo";
 import { SelectPicture } from "../mini/SelectPicture";
