+++ packages/screens/FeedNewArticle/FeedNewArticleScreen.tsx
@@ -11,8 +11,8 @@ import { SVG } from "@/components/SVG";
 import { ScreenContainer } from "@/components/ScreenContainer";
 import { WalletStatusBox } from "@/components/WalletStatusBox";
 import { TertiaryBox } from "@/components/boxes/TertiaryBox";
-import { FileUploader } from "@/components/fileUploader";
 import { Label, TextInputCustom } from "@/components/inputs/TextInputCustom";
+import { FileUploader } from "@/components/inputs/fileUploader";
 import { FeedPostingProgressBar } from "@/components/loaders/FeedPostingProgressBar";
 import { RichText } from "@/components/socialFeed/RichText";
 import { PublishValues } from "@/components/socialFeed/RichText/RichText.type";
