+++ packages/screens/Projects/ProjectsMakeRequestScreen/ShortPresentation.tsx
@@ -7,9 +7,9 @@ import { MakeRequestFooter } from "./Footer";
 import addSVG from "../../../../assets/icons/add.svg";
 import { BrandText } from "../../../components/BrandText";
 import { PrimaryButtonOutline } from "../../../components/buttons/PrimaryButtonOutline";
-import { FileUploader } from "../../../components/fileUploader";
 import { RoundedGradientImage } from "../../../components/images/RoundedGradientImage";
 import { TextInputCustom } from "../../../components/inputs/TextInputCustom";
+import { FileUploader } from "../../../components/inputs/fileUploader";
 import { SpacerColumn } from "../../../components/spacer";
 import { useNameSearch } from "../../../hooks/search/useNameSearch";
 import { useSelectedNetworkId } from "../../../hooks/useSelectedNetwork";
