+++ packages/screens/UserPublicProfile/components/mintUploader/mintUploader.web.tsx
@@ -9,8 +9,8 @@ import { DeleteButton } from "@/components/FilePreview/DeleteButton";
 import { OptimizedImage } from "@/components/OptimizedImage";
 import { SVG } from "@/components/SVG";
 import { PrimaryBox } from "@/components/boxes/PrimaryBox";
-import { FileUploaderProps } from "@/components/fileUploader/FileUploader.type";
-import { formatFile } from "@/components/fileUploader/formatFile";
+import { FileUploaderProps } from "@/components/inputs/fileUploader/FileUploader.type";
+import { formatFile } from "@/components/inputs/fileUploader/formatFile";
 import { AnimatedLoader } from "@/components/loaders/AnimatedLoader";
 import { SpacerColumn } from "@/components/spacer";
 import { useFeedbacks } from "@/context/FeedbacksProvider";
